requirements-serve.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.0.0
python-multipart